separate the code in PinterestLayout in different files.
