+++ Papr/Custom UI/PinterestLayout.swift
@@ -92,45 +92,4 @@ class PinterestLayout: UICollectionViewLayout {
     }
 }
 
-extension Reactive where Base: PinterestLayout {
-
-    var delegate: RxPinterestLayoutDelegateProxy {
-        return RxPinterestLayoutDelegateProxy.proxy(for: base)
-    }
-    
-    func updateSize(_ indexPath: IndexPath) -> Binder<CGSize> {
-        return Binder(base) { base, size in
-            self.delegate.sizes[indexPath] = size
-            base.invalidateLayout()
-        }
-    }
-}
-
-class RxPinterestLayoutDelegateProxy: DelegateProxy<PinterestLayout, PinterestLayoutDelegate>, PinterestLayoutDelegate, DelegateProxyType {
-    
-    weak private(set) var pinterestLayout: PinterestLayout?
-    var sizes:[IndexPath: CGSize] = [:]
-
-    init(pinterestLayout: PinterestLayout) {
-        self.pinterestLayout = pinterestLayout
-        super.init(parentObject: pinterestLayout, delegateProxy: RxPinterestLayoutDelegateProxy.self)
-    }
-
-    static func registerKnownImplementations() {
-        self.register { RxPinterestLayoutDelegateProxy(pinterestLayout: $0) }
-    }
-
-    static func currentDelegate(for object: PinterestLayout) -> PinterestLayoutDelegate? {
-        return object.delegate
-    }
-
-    static func setCurrentDelegate(_ delegate: PinterestLayoutDelegate?, to object: PinterestLayout) {
-        object.delegate = delegate
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, sizeForPhotoAtIndexPath indexPath: IndexPath) -> CGSize {
-        return sizes[indexPath] ?? CGSize(width: 100, height: 100)
-    }
-}
-
